Rename implementation modules
